Separate the AsyncSearchSelect component out to its own file for clarity
